Update README.md to contain a warning
I thought I was clever: instead of creating a custom HttpClientHandler I instead created a custom HttpClient. I assumed that overriding HttpClient.SendAsync would accomplish the same result. I was clearly wrong. 

I added a note to the proper section of README.md to help other people avoid my mistake. If you think this warning is valuable then I'd be happy to iterate on the exact wording of the message.