Add support for content headers on empty POST requests
When we have custom headers added to the request, we now make sure there's a non-null `Content` to attach content headers to.

Because setting a body on a GET or HEAD request will blow up at runtime (`HttpClient` won't allow it), we have to make sure the request isn't one of those types before auto-adding `HttpContent`.