Move PushStreamContent Namespace
Previous behavior caused issues of mbiguous reference around
PushStreamContent when both Refit and System.Net.Http.Formatting are
included in a project. This is essentially a show stopper to upgrading
to the latest Refit.

The solution is to change the internal implementation of PushStreamContent
namespace from System.Net.Http -> Refit.Net.Http and to update
the usages internally. Note this may be a breaking change to those who
have taken a dependency on PushStreamContent in their own projects from
the Refit assemblies.