Relax Json.NET version requirements
This change downgrades the Newtonsoft.Json package reference to a minimum of 9.01. We're not using any of the newer APIs, and all our tests pass, so I think this is okay. :rocket:

It required adding an explicit dependency on `System.Runtime.Serialization.Formatters` to the netstandard 1.4 version, and in the process of adding that I cleaned up the 1.4 specific package references in the project a tiny bit.

My only concern here is the [fairly large number of fixes in Json.NET between 9.0.1 and 12.0.2](https://github.com/JamesNK/Newtonsoft.Json/releases). This change won't stop stop anyone from referencing a later version though -- explicitly or otherwise. It just might open us up to some annoying bug reports. (As stated in the referenced issue, most people will be using a later version than 9.0.1 anyway.)

So this fixes #670, I think :joy: