Reintroduced parameterless RefitSettings constructor
Made this change so that the constructor with an explicit IContentSerializer parameter doesn't directly reference the NewtonsoftJsonContentSerializer type. Hopefully this should help AOT compilers (eg. .NET Native) to correctly identify that type as an unused dependency in case the default constructor is never used, and to remove it from the compiler binaries during the dependency reduction build pass.