Fixed a bug in the readme (#1366)
A bug report came in via an old issue https://github.com/reactiveui/refit/issues/801 that highlighted a bug in the docs. I've fixed the doc bug and added some guidance on how better to support a common use case and outlined in what case it's appropriate to pass the `Polly.Context` via the `[Property]` attribute.